% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seismodith.R
\name{eqldens}
\alias{eqldens}
\title{Map magnitudes to a reference distribution}
\usage{
eqldens(M, Mref)
}
\arguments{
\item{M}{numeric; the magnitudes to map}

\item{Mref}{numeric; the reference magnitudes used to create the mapping}
}
\value{
An object with class \code{'eqldensmap'}
}
\description{
Map magnitudes to a reference distribution
}
\examples{
M.ref <- runif(10,5,8)
eqldens(1:10, M.ref)
}
\seealso{
\code{\link{seismodith}}
}

