% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{getScale}
\alias{getScale}
\alias{getScale.eqldensmap}
\title{Returns empirical cumulative distribution mappings}
\usage{
getScale(x, ...)

\method{getScale}{eqldensmap}(x, at = NULL, inverse = FALSE, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters to \code{\link{pretty}}}

\item{at}{numeric; the magnitudes to return the mapping for. If \code{NULL} (the default)
the function uses \code{\link{pretty}} to generate these.}

\item{inverse}{logical; should \code{at} be values of the ecdf? If \code{TRUE}
the ecdf is inverted to solve for the magnitudes}
}
\value{
a \code{data.frame} with the scale and mapping information
}
\description{
Returns empirical cumulative distribution mappings
}
\examples{
M.ref <- runif(10,5,8)
M.map <- eqldens(1:10, M.ref)
getScale(M.map)
}

