% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.eqldensmap}
\alias{plot.eqldensmap}
\title{Plot of the magnitude mapping}
\usage{
\method{plot}{eqldensmap}(x, ...)
}
\arguments{
\item{x}{\code{'eqldensmap'} object}

\item{...}{additional parameters to \code{\link{plot}}}
}
\description{
Plot of the magnitude mapping
}
\examples{
M.ref <- runif(10,5,8)
plot(eqldens(1:10, M.ref))
}
\seealso{
\code{\link{eqldens}}
}

