% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seismodith.R
\name{seismodith}
\alias{seismodith}
\title{Dither earthquake magnitudes assuming a Gutenberg-Richter distribution}
\usage{
seismodith(M, dither.by = 0.1, b.value = 1, offset = 0)
}
\arguments{
\item{M}{numeric; earthquake magnitude}

\item{dither.by}{numeric; the amoung to dither by}

\item{b.value}{numeric; the 'b-value' in the Gutenberg-Richter distribution}

\item{offset}{numeric; a constant to add to all magnitudes}
}
\value{
a \code{data.frame} with the original magnitudes, the dithered
magnitudes, and the uniform distribution used in the dithering
}
\description{
Dither earthquake magnitudes assuming a Gutenberg-Richter distribution
}
\examples{
1
}
\seealso{
\code{\link{eqldens}}
}

